+++ template-parts/blocks/multi-column/template.php
@@ -33,12 +33,14 @@
                     $button         = get_sub_field('button');
                     $button_color   = get_sub_field('button_color');
                     $image          = get_sub_field('image');
-                    if ( $c_count < 4 ) {
+                    if ( get_field('number_of_columns') == 2 ) {
+                        $image      = wp_get_attachment_image_src( $image, 'column_image_two' );
+                    } elseif ( get_field('number_of_columns') == 3 ) {
                         $image      = wp_get_attachment_image_src( $image, 'column_image_three' );
-                    } else {
+                    } elseif ( get_field('number_of_columns') == 4 ) {
                         $image      = wp_get_attachment_image_src( $image, 'column_image_four' );
                     }
-                    $icon       = get_sub_field('icon');
+                    $icon           = get_sub_field('icon');
                 ?>
                 <div class="<?php echo $columns; ?>">
                 <?php if ( $button && $link_column ) : ?> <a href="<?php echo $button['url']; ?>"><?php endif; ?>